sus_requirements.txt
auto-sklearn==0.15.0
scikit-learn==0.24.2
nltk==3.9.1
numpy
python-crfsuite==0.9.11
pandas==2.2.3
tensorflow==2.1.9
tqdm
hpsklearn==0.1.0
#USE PYTHON 3.9 :)